Move projection SQL query to Projection model
